Tab 틀어짐 Tab Width 계산 해서 로그 출력 - Tab 크기 이상 시 초기화 예정 Tab Pitch  계산 시 분해능 가져오기

+++ NotchingGradeInsp/CCounterQueueCtrl.cpp
@@ -79,7 +79,7 @@ BOOL CCounterQueueCtrl::FindTabId(int TabId)
 	BOOL b = FALSE;
 	::EnterCriticalSection(&m_csQueue);
 	CCounterInfo CntInfo;
-	int CntInfosize = m_CntInfo.size();
+	int CntInfosize = (int)m_CntInfo.size();
 	int loop = 0;
 	while (loop < CntInfosize)
 	{
